Implement VM with bytecode execution, native dispatch, and module system (Phase 6)
Adds the complete virtual machine including:
- Module system with ScriptFunction, NativeFunction, and NativeCallContext
- Stack-based VM with frame pointer, call stack, and 64-bit register
- Full execution loop covering ~170 opcodes: arithmetic (int/float/double/int64),
  bitwise, comparison, type conversion, control flow, and power operations
- Native function dispatch via callback interface
- Fix 1-dword instruction encoding overlap in bytecode builder (w_arg0 truncated to 8 bits)

All 195 tests pass.

+++ crates/angelscript-compiler/src/compiler/bytecode_builder.rs
@@ -185,9 +185,10 @@ fn encode_instruction(inst: &Instruction, buf: &mut Vec<u32>) {
     let size = inst.op.size_dwords();
     match size {
         1 => {
-            // Pack: opcode in low byte, word args in upper bytes.
+            // Pack: opcode in bits 0-7, w_arg0 in bits 8-15, w_arg1 in bits 16-31.
+            // w_arg0 is truncated to 8 bits (sufficient for stack offsets < 256).
             let word = (inst.op as u32)
-                | ((inst.w_arg0 as u16 as u32) << 8)
+                | (((inst.w_arg0 as u8) as u32) << 8)
                 | ((inst.w_arg1 as u16 as u32) << 16);
             buf.push(word);
         }
